requirements.txt
astroid==2.15.4
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
docutils==0.19
first==2.0.2
idna==3.4
importlib-metadata==6.6.0
isort==5.12.0
jaraco.classes==3.2.3
keyring==23.13.1
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
platformdirs==3.5.0
pydantic==1.10.7
Pygments==2.15.1
pylint==2.17.3
PyYAML==6.0
readme-renderer==37.3
requests==2.30.0
requests-toolbelt==1.0.0
responses==0.23.1
rfc3986==2.0.0
rich==13.3.5
six==1.16.0
toml==0.10.2
tomlkit==0.11.8
twine==4.0.2
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
urllib3==2.0.2
webencodings==0.5.1
wrapt==1.15.0
zipp==3.15.0
